refactored export types of components
