+++ src/runtime/server/components/drawHorizontalLine.ts
@@ -1,0 +1,14 @@
+import type { PDFDocumentType } from "../../types"
+
+/**
+  * Draw a Horizontal line across the document
+  * @param moveDown The amount of lines to move down before and after drawing the line. Default: 1
+*/
+export function drawHorizontalLine<T>(doc: PDFDocumentType<T>, moveDown = 1): void {
+  doc
+    .moveDown(moveDown)
+    .moveTo(doc.options.margins.left, doc.y)
+    .lineTo(doc.page.width - doc.options.margins.right, doc.y)
+    .stroke()
+    .moveDown(moveDown + 0.5)
+}
