+++ src/runtime/server/index.ts
@@ -1,2 +1,4 @@
 export { createPDF, streamReturnPDF } from './pdf'
+export * from './components'
+
 export type { PDFDocumentType, PDFOptions } from '../types'
