+++ src/runtime/types.ts
@@ -1,6 +1,5 @@
 import PDFDocument from 'pdfkit'
-import type { HorizontalLine } from './server/components/line'
-import type { ApplyLayout } from './server/components/layout'
+import type { ApplyLayout } from './server/layout'
 import type { LayoutOptions } from './server/pdf'
 
 export interface PDFOptions extends PDFKit.PDFDocumentOptions {
@@ -22,11 +21,5 @@ export interface ModuleOptions {
 export type PDFDocumentType<TData> = PDFDocument & {
   data?: TData
   layout?: LayoutOptions
-
-  /**
-   * Draw a Horizontal line across the document
-   * @param moveDown The amount of lines to move down before and after drawing the line. Default: 1
-   */
-  horizontalLine: HorizontalLine
   applyLayout: ApplyLayout
 }
